requirements.txt
google-genai>=1.0.0
jinja2>=3.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
